requirements.txt
streamlit==0.81.0
pandas==1.2.4
mysql-connector-python==8.0.24
nltk==3.6.2
plotly==4.14.3
matplotlib==3.4.1
wordcloud==1.8.1
spacy==3.0.6
discord.py==1.7.2
python-dotenv==0.17.1
better-profanity==0.7.0
PyYAML==5.4.1
asyncio
setuptools
wheel